Refactor alert-related imports to streamline hook usage by consolidating alert query hooks into a single file. Introduce new alert service for improved alert management functionality, including create, update, and delete operations.
